+++ src/main/java/io/ticticboom/mods/mm/setup/MMRegisters.java
@@ -3,6 +3,8 @@ package io.ticticboom.mods.mm.setup;
 import io.ticticboom.mods.mm.Ref;
 import io.ticticboom.mods.mm.debug.tool.DebugToolItem;
 import io.ticticboom.mods.mm.item.BlueprintItem;
+import io.ticticboom.mods.mm.item.MultiblockSaverItem;
+import io.ticticboom.mods.mm.item.PrioritySetterItem;
 import io.ticticboom.mods.mm.structure.StructureManager;
 import net.minecraft.core.registries.Registries;
 import net.minecraft.network.chat.Component;
@@ -13,6 +15,7 @@ import net.minecraft.world.level.block.entity.BlockEntityType;
 import net.minecraftforge.fml.javafmlmod.FMLJavaModLoadingContext;
 import net.minecraftforge.registries.*;
 
+@SuppressWarnings("unused")
 public class MMRegisters {
     public static final DeferredRegister<Block> BLOCKS = DeferredRegister.create(ForgeRegistries.BLOCKS, Ref.ID);
     public static final DeferredRegister<Item> ITEMS = DeferredRegister.create(ForgeRegistries.ITEMS, Ref.ID);
@@ -22,7 +25,8 @@ public class MMRegisters {
 
     public static final RegistryObject<BlueprintItem> BLUEPRINT = ITEMS.register("blueprint", BlueprintItem::new);
     public static final RegistryObject<Item> DEBUG_TOOL = ITEMS.register("debug_tool", DebugToolItem::new);
-    public static final RegistryObject<Item> PRIORITY_SETTER = ITEMS.register("priority_setter", () -> new io.ticticboom.mods.mm.item.PrioritySetterItem());
+    public static final RegistryObject<Item> PRIORITY_SETTER = ITEMS.register("priority_setter", PrioritySetterItem::new);
+    public static final RegistryObject<MultiblockSaverItem> MULTIBLOCK_SAVER = ITEMS.register("multiblock_saver", MultiblockSaverItem::new);
 
     public static final RegistryObject<CreativeModeTab> MM_TAB = TABS.register("mm", () -> CreativeModeTab.builder().title(Component.translatable("tab.mm.main"))
             .icon(() -> BLUEPRINT.get().getDefaultInstance())
@@ -38,7 +42,6 @@ public class MMRegisters {
             .displayItems((p, o) -> o.acceptAll(StructureManager.STRUCTURE_BLUEPRINTS.values()))
             .build());
 
-    @SuppressWarnings("deprecation")
     public static void register() {
         @SuppressWarnings("removal") var bus = FMLJavaModLoadingContext.get().getModEventBus();
         BLOCKS.register(bus);